+++ DeliveryRun/DeliveryRun iOS/RobbyViewController.swift
@@ -60,17 +60,3 @@ class RobbyViewController: UIViewController {
     }
 }
 
-
-class CustomGameButton: UIButton {
-    required init (coder aDecoder:NSCoder) {
-        super.init(coder: aDecoder)!
-        self.backgroundColor = .deliveryrunBlack
-        self.titleLabel?.font = UIFont(name: "BMJUAOTF", size: 30)
-        self.tintColor = .white
-        self.setPreferredSymbolConfiguration(UIImage.SymbolConfiguration(pointSize: 30), forImageIn: .normal)
-        self.layer.borderColor = UIColor.white.cgColor
-        self.layer.borderWidth = 2
-        self.layer.cornerRadius = 10
-    }
-}
-
